+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Unarmored Movement now works correctly (again) with Active Effect generation when using DAE SRD. (@miromonti)
 * Midi-SRD module name was incorrect during module detection.
+* Healing Hands feature - better support.
 
 # 2.3.2